Implement EEPROM reset firmware

+++ source/ATmegaEEPROM/src/main.cpp
@@ -14,13 +14,12 @@
 #endif
 
 // 3rd-Party Libraries
-#include <CRC32.h>
 #include <EEPROM.h>
 
 // Function: clearEEPROM
 // Purpose: Clears the entire EEPROM by writing 0xFF to all addresses.
 void clearEEPROM() {
-  for (int i = 0; i < EEPROM.length(); i++) {
+  for (uint16_t i = 0; i < EEPROM.length(); i++) {
     EEPROM.write(i, 0xFF); // Write 0xFF to each address
   }
   debugln(F("EEPROM cleared."));
@@ -30,7 +29,7 @@ void clearEEPROM() {
 // Purpose: Checks if the entire EEPROM is cleared (all bytes are 0xFF).
 // Returns: true if cleared, false otherwise.
 bool isEEPROMCleared() {
-  for (int i = 0; i < EEPROM.length(); i++) {
+  for (uint16_t i = 0; i < EEPROM.length(); i++) {
     if (EEPROM.read(i) != 0xFF) {
       return false; // Found a non-cleared byte
     }
